+++ pyanatem/__init__.py
@@ -64,4 +64,4 @@ __all__ = [
     "ResultadoSAV",
 ]
 
-__version__ = "1.1.2"
+__version__ = "1.1.3"